% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{arrows}
\alias{arrows}
\title{Add styled arrows to a plot}
\usage{
arrows(x0, y0, x1 = x0, y1 = y0, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x0, y0}{coordinates of *from* endpoint}

\item{x1, y1}{coordinates of *to* endpoint}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStle()}

\item{Rcssclass}{character, style class}

\item{...}{Futher parameters, see documentation of graphics::arrows}
}
\description{
This is a wrapper for R's arrows function.
See R's documentation for graphics::arrows for further details.
}
\examples{

# draw an arrow
plot(c(0, 1), c(0, 1), type="n")
arrows(0.2, 0.2, x1=0.8, y1=0.5)

}
