% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{barplot}
\alias{barplot}
\title{Draw a styled barplot}
\usage{
barplot(height, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{height}{numeric vector giving bar lengths}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of graphics::barplot}
}
\description{
This is a wrapper for R's barplot function.
See R's documentation for graphics::barplot for further details.
}
