% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{boxplot}
\alias{boxplot}
\title{Draw a styled boxplot}
\usage{
boxplot(x, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x}{data for boxplot; either single numeric vector or a list
of numeric vectors; see documentation of boxplot()}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{character, style class}

\item{...}{Further parameters, see documentation of graphics::boxplot}
}
\description{
This is a wrapper for R's boxplot function.
See R's documentation for graphics::boxplot for further details.
}
\examples{

# draw a complete boxplot
dataset <- list(A=rpois(30, 10), B=rpois(30, 20))
boxplot(dataset, col=c("#dd0000", "#dd8888"))

}
