% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{cairo_pdf}
\alias{cairo_pdf}
\title{Create a styled cairo_pdf figure}
\usage{
cairo_pdf(filename, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{filename}{character string with file name}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{character, style class}

\item{...}{Further parameters, see documentation of grDevices::pdf}
}
\description{
This is a wrapper for R's cairo_pdf function.
See R's documentation for grDevices::cairo_pdf for further details
}
\details{
Note this uses styles from 'pdf' css selectors
}
\examples{

# send content of graphics to a pdf file
# to run this, un-comment the pdf() and dev.off() lines
# cairo_pdf(file="example-file.pdf")
barplot(1:5)
# dev.off()

}
