% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssMisc.R
\name{ctext}
\alias{ctext}
\title{Write styled text into a plot corner}
\usage{
ctext(
  label,
  x = NULL,
  y = NULL,
  adj = NULL,
  cex = NULL,
  Rcss = "default",
  Rcssclass = NULL,
  ...
)
}
\arguments{
\item{label}{character, text for corner label}

\item{x, y}{numeric, positions for manual placement}

\item{adj}{numeric of length 2, argument adj for text}

\item{cex}{numeric, argument cex for text}

\item{Rcss}{style sheet object}

\item{Rcssclass}{character, style class}

\item{...}{additional argument, passed to text()}
}
\description{
This can be suitable for placing a label in a multi-panel figure.
Note the automatic placement does not work when a plot is generated
with logarithmic scales.
}
\examples{

plot(1:10, 1:10)
ctext("A")

}
