% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{RcssSetDefaultStyle}
\alias{RcssGetDefaultClass}
\alias{RcssGetDefaultStyle}
\alias{RcssSetDefaultClass}
\alias{RcssSetDefaultStyle}
\title{Get/Set Default Rcssplot style sheet}
\usage{
RcssSetDefaultStyle(Rcss)

RcssGetDefaultStyle()

RcssSetDefaultClass(class)

RcssGetDefaultClass()
}
\arguments{
\item{Rcss}{style sheet object (set NULL to remove default style)}

\item{class}{the default class}
}
\description{
Set a default Rcssplot style sheet. This style sheet will be
applied in all functions of the Rcss family. See also related
function RcssGetDefaultStyle()
}

