% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{dryer}
\alias{dryer}
\title{RCSS - Don't repeat yourself}
\usage{
dryer(Rcss)
}
\arguments{
\item{Rcss}{the style sheet to reuse}
}
\value{
a 'convert to Rcss' object
}
\description{
RCSS - Don't repeat yourself
}

