% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssGet.R
\name{getAllProperties}
\alias{getAllProperties}
\title{get properties, recursively}
\usage{
getAllProperties(RcssProperties)
}
\arguments{
\item{RcssProperties}{object}
}
\value{
character vector
}
\description{
get properties, recursively
}
\keyword{internal}
