% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{grid}
\alias{grid}
\title{#' Draw a styled grid}
\usage{
grid(Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of graphics::grid}
}
\description{
This is a wrapper for R's grid function.
See R's documentation for graphics::grid for further details.
}
