% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{matplot}
\alias{matplot}
\title{Add styled line segments to a plot}
\usage{
matplot(x, y, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x, y}{vectors or matrices of data for plotting. The number of rows 
should match. If one of them are missing, the other is taken as y and an x 
vector of 1:n is used. Missing values (NAs) are allowed.}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{character, style class}

\item{...}{Further parameters, see documentation of graphics::lines}
}
\description{
This is a wrapper for R's matplot function.
See R's documentation for graphics::matplot for further details.
}
\examples{

# draw scatter based on column in a matrix
dataset = cbind(A=rnorm(20), B=rnorm(20))
matplot(dataset)

}
