% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{mtext}
\alias{mtext}
\title{Write styled text into a plot margin}
\usage{
mtext(text, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{text}{characters to print on the plot}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of graphics::mtext}
}
\description{
This is a wrapper for R's mtext function.
See R's documentation for graphics::mtext for further details.
}
\examples{

# draw text into a margin
plot(c(0, 1), c(0, 1), type="n", xlab="", ylab="")
mtext(side=1, "bottom x-axis label", line=2.5)
mtext(side=2, "left y-axis label", line=2.5)
mtext(side=3, "top x-axis label")
mtext(side=4, "right y-axis label")

}
