% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssMisc.R
\name{parplot}
\alias{parplot}
\title{combination of par and plot}
\usage{
parplot(x, y, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x, y}{coordinates for points on the plot}

\item{Rcss}{style sheet object, leave "default" to use a style
defined via RcssDefaultStyle()}

\item{Rcssclass}{character, style class}

\item{...}{Further parameters, passed to plot()}
}
\description{
The sequence of par() and plot() occurs so frequently that
it a shortcut is helpful.
}
\examples{
parplot(x=1:4, y=c(1,3,2,4))
}
