% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssParser.R
\name{parseIDENT}
\alias{parseIDENT}
\title{parse special strings into R primitives}
\usage{
parseIDENT(tok)
}
\arguments{
\item{tok}{one string}
}
\value{
R primitive for that string, "NULL" gives an empty list
}
\description{
parse special strings into R primitives
}
