% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{pdf}
\alias{pdf}
\title{Create a styled pdf figure}
\usage{
pdf(file, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{file}{character string with file name}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of grDevices::pdf}
}
\description{
This is a wrapper for R's pdf function.
See R's documentation for grDevices::pdf for further details
}
