% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{png}
\alias{png}
\title{Create a styled png figure}
\usage{
png(file, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{file}{character string with file name}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of grDevices::png}
}
\description{
This is a wrapper for R's png function.
See R's documentation for grDevices::png for further details.
}
