% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{points}
\alias{points}
\title{Add styled points to a plot}
\usage{
points(x, y = NULL, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x, y}{coordinates for points on the plot}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of graphics::points}
}
\description{
This is a wrapper for R's points function.
See R's documentation for graphics::points for further details.
}
