% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rcss.R
\name{print.Rcss}
\alias{print.Rcss}
\title{Show basic information about an Rcss object}
\usage{
\method{print}{Rcss}(x, ...)
}
\arguments{
\item{x}{style sheet object}

\item{...}{Further parameters are ignored}
}
\description{
Display selectors encoded in an Rcss object.
For more detailed information about the object, see function printRcss()
}

