% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcssarrows}
\alias{Rcssarrows}
\title{Add styled arrows to a plot}
\usage{
Rcssarrows(x0, y0, x1 = x0, y1 = y0, Rcss = "default", Rcssclass = NULL,
  ...)
}
\arguments{
\item{x0,y0}{coordinates of *from* endpoint}

\item{x1,y1}{coordinates of *to* endpoint}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Futher parameters, see documentation of arrows()}
}
\description{
Rcssarrows is a wrapper for R's arrows() function.
See R's documentation for arrows() for further details.
}

