% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcssbarplot}
\alias{Rcssbarplot}
\title{Draw a styled barplot}
\usage{
Rcssbarplot(height, Rcss = NULL, Rcssclass = NULL, ...)
}
\arguments{
\item{height}{numeric vector giving bar lengths}

\item{Rcss}{style sheet object}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of barplot()}
}
\description{
Rcssbarplot is a wrapper for R's barplot() function.
See R's documentation for barplot() for further details.
}

