% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcssbox}
\alias{Rcssbox}
\title{Add a styled box around a plot}
\usage{
Rcssbox(which = "plot", Rcss = NULL, Rcssclass = NULL, ...)
}
\arguments{
\item{which}{character specifying where to draw a box;
see documentation of box()}

\item{Rcss}{style sheet object}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of box()}
}
\description{
Rcssbox is a wrapper for R's box() function.
See R's documentation for box() for further details.
}

