% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcssboxplot}
\alias{Rcssboxplot}
\title{Draw a styled boxplot}
\usage{
Rcssboxplot(x, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x}{data for boxplot; either single numeric vector or a list
of numeric vectors; see documentation of boxplot()}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of boxplot()}
}
\description{
Rcssboxplot is a wrapper for R's boxplot() function.
See R's documentation for boxplot() for further details.
}
