% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssChangePropertyValue}
\alias{RcssChangePropertyValue}
\title{Modify an Rcss style sheet object}
\usage{
RcssChangePropertyValue(Rcss, selector, Rcssclass = NULL,
  propertylist = NULL, property = NULL, value = NULL)
}
\arguments{
\item{Rcss}{style sheet object}

\item{selector}{name of one selector ("text", "plot", "axis", etc.)}

\item{Rcssclass}{subclass of style sheet. Leave NULL to change
base property. Provide one character value to edit one subclass.
Provide a vector to edit a subclass of a subclass of a ...}

\item{propertylist}{list with property/value pairs to update}

\item{property}{name of a single property. This is only used
when propertylist is set to NULL}

\item{value}{new values associated with property above. This is
only used propertylist is set to NULL}
}
\description{
Creates a new Rcss style sheet object from the input, modifying
one or more properties.
}

