% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\docType{data}
\name{RcssCompulsoryClass}
\alias{RcssCompulsoryClass}
\title{Vector holding set a compulsory Rcssclass}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
RcssCompulsoryClass
}
\description{
These style class (or classes) are applied in all functions of
the Rcss family.
}
\keyword{datasets}
