% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssConstructor}
\alias{RcssConstructor}
\title{creates an empty Rcss object (i.e. just a list with a class name)
the items inserted into this list/object will be called "selectors"}
\usage{
RcssConstructor()
}
\description{
creates an empty Rcss object (i.e. just a list with a class name)
the items inserted into this list/object will be called "selectors"
}
