% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcsscontour}
\alias{Rcsscontour}
\title{Draw a styled contour}
\usage{
Rcsscontour(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1, length.out =
  ncol(z)), z, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x}{numeric vector; locations of grid lines}

\item{y}{numeric vector; locations of grid lines}

\item{z}{matrix of values}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of contour()}
}
\description{
Rcsscontour is a wrapper for R's contour() function.
See R's documentation for contour() for further details.
}
