% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\docType{data}
\name{RcssDefaultStyle}
\alias{RcssDefaultStyle}
\title{Default Rcssplot style sheet}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
RcssDefaultStyle
}
\description{
This style sheet will be applied in all functions of the Rcss family.
}
\keyword{datasets}
