% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssLexer.R
\name{RcssFileCheckRead}
\alias{RcssFileCheckRead}
\title{check if a file exists, then read it}
\usage{
RcssFileCheckRead(f)
}
\arguments{
\item{f}{character, single file name}
}
\value{
contents of file with " \n" at end of each line added
(the space-newline is convenient for lexing)
}
\description{
check if a file exists, then read it
}
