% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssGetDefault}
\alias{RcssGetDefault}
\title{gets an object from a calling environment (recursively)}
\usage{
RcssGetDefault(what)
}
\arguments{
\item{what}{character, use either "RcssDefaultStyle" or "RcssCompulsoryClass"}
}
\description{
note: automatic object fetching uses chain of binding environments,
while this uses parent.frame() chain
}
