% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssGetDefaultStyle}
\alias{RcssGetDefaultStyle}
\title{Get default Rcssplot style object}
\usage{
RcssGetDefaultStyle(Rcss = "default")
}
\arguments{
\item{Rcss}{Rcss object, replacement default style object. When
set to "default", the function returns a copy of the default object
defined in parent environment. When set to Rcss object, the function
ignores the default and returns the set object back.}
}
\description{
Fetches the value of the RcssDefaultStyle object defined in
parent environments.
}

