% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssGet.R
\name{RcssGetProperties}
\alias{RcssGetProperties}
\title{get a list of properties relevant for a given selector and Rcssclasses
(Used internally, 2 frames down from user interaction)}
\usage{
RcssGetProperties(Rcss, selector, Rcssclass = NULL)
}
\arguments{
\item{Rcss}{object}

\item{selector}{character}

\item{Rcssclass}{character, style class}
}
\description{
get a list of properties relevant for a given selector and Rcssclasses
(Used internally, 2 frames down from user interaction)
}
\keyword{internal}
