% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssLexer.R
\name{RcssGetToken}
\alias{RcssGetToken}
\title{concatenate a subset of characters}
\usage{
RcssGetToken(cc, pos, newpos)
}
\arguments{
\item{cc}{character vector}

\item{pos}{integer, start position}

\item{newpos}{integer, end position}
}
\value{
a string containing characters in the interval
}
\description{
concatenate a subset of characters
}
\keyword{internal}
