% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcssgrid}
\alias{Rcssgrid}
\title{##' Draw a styled grid}
\usage{
Rcssgrid(Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of grid()}
}
\description{
Rcssgrid is a wrapper for R's \code{\link{grid}}() function.
See R's documentation for \code{\link{grid}}() for further details.
}
