% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssGet.R
\name{RcssHelperGetPropertyValue}
\alias{RcssHelperGetPropertyValue}
\title{traverses the style sheets breadth first and obtains values for the property}
\usage{
RcssHelperGetPropertyValue(RcssProperties, property, Rcssclass = NULL,
  inclass = c(), bestvalue = NULL)
}
\arguments{
\item{RcssProperties}{object}

\item{property}{character}

\item{Rcssclass}{character, style class}

\item{inclass}{vector of class names (used in recursion)}

\item{bestvalue}{list with best-guess property value}
}
\value{
list(level, values)
(From this, another function can choose between competing values)
}
\description{
This is a helper function for RcssGetPropertyValue
}
