% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcsshist}
\alias{Rcsshist}
\title{##' Draw a styled histogram}
\usage{
Rcsshist(x, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x}{numeric vector}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of hist()}
}
\description{
Rcsshist is a wrapper for R's hist() function.
See R's documentation for hist() for further details.
}

