% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssLexer.R
\name{RcssIsEscaped}
\alias{RcssIsEscaped}
\title{helper for string parsing}
\usage{
RcssIsEscaped(cc, pos)
}
\arguments{
\item{cc}{- vector of characters}

\item{nowpos}{- position of character to check escape for (e.g. a " in a string)}
}
\value{
true if position  is preceded by an odd number of slashes
}
\description{
helper for string parsing
}
