% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcssjpeg}
\alias{Rcssjpeg}
\title{Create a styled jpg figure}
\usage{
Rcssjpeg(file, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{file}{character string with file name}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of jpeg()}
}
\description{
Rcssjpeg is a wrapper for R's jpeg() function.
See R's documentation for jpeg() for further details
}

