% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcsslegend}
\alias{Rcsslegend}
\title{Add a styled legend to aplot}
\usage{
Rcsslegend(x, y = NULL, legend, Rcss = NULL, Rcssclass = NULL, ...)
}
\arguments{
\item{x,y}{position of the legend}

\item{legend}{character vector with labels (text appears in the legend)}

\item{Rcss}{style sheet object}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of legend()}
}
\description{
Rcsslegend is a wrapper for R's legend() function.
See R's documentation for legend() for further details.
}

