% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssLexer.R
\name{RcssLexer}
\alias{RcssLexer}
\title{start the Lexer}
\usage{
RcssLexer(file = NULL, text = NULL)
}
\arguments{
\item{text}{character, strings with Rcss}

\item{f}{character, input file, or vector of input files}
}
\value{
data frame with tokens
}
\description{
start the Lexer
}
\keyword{internal}
