% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssLexer.R
\name{RcssLexNextToken}
\alias{RcssLexNextToken}
\title{process one token starting at position 'pos'}
\usage{
RcssLexNextToken(cc, pos, terminals = unlist(strsplit(".,;:{}=", "")),
  spacechars = unlist(strsplit(" \\t\\r\\n\\f", "")))
}
\arguments{
\item{cc}{vector of characters}

\item{pos}{integer, current position in the cc vector}

\item{terminals}{character vector, characters that mark end of a token}

\item{spacechars}{character vector, all characters interpreted as a space}
}
\value{
triple (NEWPOSITION, TOKEN, TOKENTYPE)
}
\description{
process one token starting at position 'pos'
}
