% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcsslines}
\alias{Rcsslines}
\title{Add styled line segments to a plot}
\usage{
Rcsslines(x, y = NULL, Rcss = NULL, Rcssclass = NULL, ...)
}
\arguments{
\item{x,y}{coordinates for start and end points for lines}

\item{Rcss}{style sheet object}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of lines()}
}
\description{
Rcsslines is a wrapper for R's lines() function.
See R's documentation for lines() for further details.
}

