% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssHelpers.R
\name{RcssMakeCallCodeString}
\alias{RcssMakeCallCodeString}
\title{produces code of the type
"col=VALUE, lty=VALUE, " (with comma at the end)}
\usage{
RcssMakeCallCodeString(varnames, listname)
}
\arguments{
\item{varnames}{character vector}

\item{listname}{character}
}
\value{
character
}
\description{
produces code of the type
"col=VALUE, lty=VALUE, " (with comma at the end)
}
