% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssParser.R
\name{RcssMakeParseTree}
\alias{RcssMakeParseTree}
\title{convert lexed tokens into a tree with selectors and declarations}
\usage{
RcssMakeParseTree(lextab)
}
\arguments{
\item{lextab}{data.frame with lexed tokens}
}
\value{
tree
}
\description{
convert lexed tokens into a tree with selectors and declarations
}
