% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcssmatplot}
\alias{Rcssmatplot}
\title{Add styled line segments to a plot}
\usage{
Rcssmatplot(x, y, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x,y}{vectors or matrices of data for plotting. The number of rows
should match. If one of them are missing, the other is taken as y and an x
vector of 1:n is used. Missing values (NAs) are allowed.}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of lines()}
}
\description{
Rcssmatplot is a wrapper for R's \code{\link{matplot}()} function.
See R's documentation for \code{\link{matplot}()} for further details.
}

