% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcssmtext}
\alias{Rcssmtext}
\title{Write styled text into a plot margin}
\usage{
Rcssmtext(text, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{text}{characters to print on the plot}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of mtext()}
}
\description{
Rcssmtext is a wrapper for R's mtext() function.
See R's documentation for mtext() for further details.
}

