% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssOverload}
\alias{RcssOverload}
\title{Overloads base graphics functions by their Rcssplot wrappers}
\usage{
RcssOverload()
}
\description{
Rcssplot graphics functions are defined using Rcss prefixes,
e.g Rcsstext(). This function can be invoked to overload the
existing functions by their Rcss wrappers. i.e. After executing
this function, the user can write execute e.g. text() and
automatically use the Rcss capabilities.
}

