% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcsspar}
\alias{Rcsspar}
\title{Set styled parameters for base graphics}
\usage{
Rcsspar(Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of lines()}
}
\description{
Rcsspar is a wrapper for R's par() function.
See R's documentation for par() for further details.
}

