% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssParser.R
\name{RcssParseDeclaration}
\alias{RcssParseDeclaration}
\title{Parse one declaration (expects an IDENT)}
\usage{
RcssParseDeclaration(lextab, n)
}
\arguments{
\item{lextab}{data frame}

\item{n}{integer, row identifier in lextab}
}
\description{
Parse one declaration (expects an IDENT)
}
\keyword{internal}
