% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssParser.R
\name{RcssParseDeclarationSet}
\alias{RcssParseDeclarationSet}
\title{parse a set of declarations
(expects to start with a '{' and end with a '}')}
\usage{
RcssParseDeclarationSet(lextab, n)
}
\arguments{
\item{lextab}{data frame with tokens}

\item{n}{integer, current row in the data.frame to process}
}
\description{
parse a set of declarations
(expects to start with a '{' and end with a '}')
}
