% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssParser.R
\name{RcssParseError}
\alias{RcssParseError}
\title{wrapper to produce generic errors}
\usage{
RcssParseError(lextab, n, parsefunction, expecting)
}
\arguments{
\item{lextab}{data frame with lexed information}

\item{n}{integer, row number in lextab}

\item{parsefunction}{character, used in error message}

\item{expecting}{character, used to hint what the expected token should have been}
}
\description{
wrapper to produce generic errors
}
