% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssLexer.R
\name{RcssParseNumber}
\alias{RcssParseNumber}
\title{parse a number}
\usage{
RcssParseNumber(chardata, pos, exponent = FALSE, decimal = FALSE)
}
\arguments{
\item{chardata}{data frame with characters and line numbers}

\item{pos}{integer, current position (expects '-' or 0-9)}

\item{exponent}{logical, set TRUE if parsing an exponent of a number}

\item{decimal}{logical, set TRUE if parsing digits after a decimal point}
}
\value{
- position of first character outside the number
}
\description{
parse a number
}
\keyword{internal}
