% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssParser.R
\name{RcssParseProperty}
\alias{RcssParseProperty}
\title{parse the name of a property}
\usage{
RcssParseProperty(lextab, n)
}
\arguments{
\item{lextab}{data frame}

\item{n}{integer, row number in lextab}
}
\description{
This is usually one token,
but if a property has dots (e.g. cex.axis) this function
will concatenate the text together
}
