% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssParser.R
\name{RcssParser}
\alias{RcssParser}
\title{parse a text file assuming it is in css format.}
\usage{
RcssParser(file)
}
\arguments{
\item{file}{character, files on disk}
}
\value{
Rcss object augmented by settings from the file(s)
}
\description{
parse a text file assuming it is in css format.
}
