% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssParser.R
\name{RcssParseRuleSet}
\alias{RcssParseRuleSet}
\title{parse one rule set (Selectors { Declarations })}
\usage{
RcssParseRuleSet(lextab, n)
}
\arguments{
\item{lextab}{data frame with tokens}

\item{n}{inteer, current row in the data frame}
}
\description{
parse one rule set (Selectors { Declarations })
}
\keyword{internal}
