% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssParser.R
\name{RcssParseSelector}
\alias{RcssParseSelector}
\title{parse one selector (IDENT [ class]* | class)}
\usage{
RcssParseSelector(lextab, n)
}
\arguments{
\item{lextab}{data frame with tokens}

\item{n}{ineger, current row in the data frame}
}
\value{
integer, new n after parsed selectors, list of selectors
}
\description{
parse one selector (IDENT [ class]* | class)
}
\keyword{internal}
