% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssParser.R
\name{RcssParseSelectorSet}
\alias{RcssParseSelectorSet}
\title{parse a set of selectors (i.e. selector [ ',' selector ]*)}
\usage{
RcssParseSelectorSet(lextab, n)
}
\arguments{
\item{lextab}{data frame with tokens}

\item{n}{integer, current row in the data frame}
}
\description{
parse a set of selectors (i.e. selector [ ',' selector ]*)
}
\keyword{internal}
