% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcsspdf}
\alias{Rcsspdf}
\title{Create a styled pdf figure}
\usage{
Rcsspdf(file, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{file}{character string with file name}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of pdf()}
}
\description{
Rcsspdf is a wrapper for R's pdf() function.
See R's documentation for pdf() for further details
}

