% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcsspng}
\alias{Rcsspng}
\title{Create a styled png figure}
\usage{
Rcsspng(file, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{file}{character string with file name}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of png()}
}
\description{
Rcsspng is a wrapper for R's png() function.
See R's documentation for png() for further details
}

