% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcsspoints}
\alias{Rcsspoints}
\title{Add styled points to a plot}
\usage{
Rcsspoints(x, y = NULL, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x, y}{coordinates for points on the plot}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of points()}
}
\description{
Rcsspoints is a wrapper for R's points() function.
See R's documentation for points() for further details.
}
