% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssUpdate.R
\name{RcssPropertiesChangeValue}
\alias{RcssPropertiesChangeValue}
\title{Helper to RcssChangePropertyValue (very similar name...)}
\usage{
RcssPropertiesChangeValue(RcssProperties, Rcssclass = NULL,
  propertylist = NULL)
}
\arguments{
\item{RcssProperties}{object}

\item{Rcssclass}{character, style class}

\item{propertylist}{list}
}
\description{
Helper to RcssChangePropertyValue (very similar name...)
}
