% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssPropertiesConstructor}
\alias{RcssPropertiesConstructor}
\title{make Rcss structure for one Rcss selector
(includes space for base properties and for subclasses)}
\usage{
RcssPropertiesConstructor()
}
\description{
make Rcss structure for one Rcss selector
(includes space for base properties and for subclasses)
}
