% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssPropertiesContainsClass}
\alias{RcssPropertiesContainsClass}
\title{object maintenance}
\usage{
RcssPropertiesContainsClass(RcssProperties, Rcssclass)
}
\arguments{
\item{RcssProperties}{object}

\item{Rcssclass}{character, style class}
}
\value{
true if the Rcssclass has been defined for all the selectors
}
\description{
object maintenance
}
