% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcssrect}
\alias{Rcssrect}
\title{Draw styled rectangles on a plot}
\usage{
Rcssrect(xleft, ybottom, xright, ytop, Rcss = "default", Rcssclass = NULL,
  ...)
}
\arguments{
\item{xleft, ybottom, xright, ytop}{vector of coordinates for rectangles'
vertices}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of rect()}
}
\description{
Rcssrect is a wrapper for R's rect() function.
See R's documentation for rect() for further details.
}

