% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssSetCompulsoryClass}
\alias{RcssSetCompulsoryClass}
\title{Set compulsory Rcssclass}
\usage{
RcssSetCompulsoryClass(Rcssclass = NA, add = TRUE)
}
\arguments{
\item{Rcssclass}{character vector, set of compulsary classes.
Setting this to c() does not affect the current value. Set to
NA to reset, i.e. remove all compulsory classes.}

\item{add}{logical, set TRUE to add Rcssclass to an existing
set of classes}
}
\description{
These style classes are applied in all functions of the Rcss family.
See also related function RcssGetCompulsoryClass()
}

