% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssSetDefaultStyle}
\alias{RcssSetDefaultStyle}
\title{Set default Rcssplot style sheet}
\usage{
RcssSetDefaultStyle(Rcss = NA)
}
\arguments{
\item{Rcss}{Rcss object, use NA to reset the default}
}
\description{
This style sheet will be applied in all functions of the Rcss family.
}

