% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcssstripchart}
\alias{Rcssstripchart}
\title{Draw styled strip chart}
\usage{
Rcssstripchart(x, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x}{list of numeric vectors}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of stripchart()}
}
\description{
Rcssstripchart is a wrapper for R's stripchart() function.
See R's documentation for stripchart() for further details.
}

