% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcsstext}
\alias{Rcsstext}
\title{Add styled text to a plot}
\usage{
Rcsstext(x, y = NULL, labels = seq_along(x), Rcss = "default",
  Rcssclass = NULL, ...)
}
\arguments{
\item{x, y}{coordinates where to write labels}

\item{labels}{characters to print on the plot}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of text()}
}
\description{
Rcsstext is a wrapper for R's text() function.
See R's documentation for text() for further details.
}

