% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcsstitle}
\alias{Rcsstitle}
\title{Add styled annotation to a plot}
\usage{
Rcsstitle(main = NULL, sub = NULL, xlab = NULL, ylab = NULL,
  Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{main}{plot title}

\item{sub}{plot sub title}

\item{xlab, ylab}{labels on axes}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of title()}
}
\description{
Rcsstitle is a wrapper for R's title() function.
See R's documentation for title() for further details.
}

