% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssUpdate.R
\name{RcssUpdateProperties}
\alias{RcssUpdateProperties}
\title{update a current set of properties and values}
\usage{
RcssUpdateProperties(nowcss, changelist)
}
\arguments{
\item{nowcss}{- a starting list of properties}

\item{changelist}{- a list of properties to update}
}
\value{
a list similar to nowcss, but with values updated according
to changelist
}
\description{
update a current set of properties and values
}
\keyword{internal}
