% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{rect}
\alias{rect}
\title{Draw styled rectangles on a plot}
\usage{
rect(xleft, ybottom, xright, ytop, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{xleft, ybottom, xright, ytop}{vector of coordinates for rectangles'
vertices}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{character, style class}

\item{...}{Further parameters, see documentation of graphics::rect}
}
\description{
This is a wrapper for R's rect function.
See R's documentation for graphics::rect for further details.
}
\examples{

# draw rectangles on an existing plot
plot(c(0, 10), c(0, 10), type="n", xlab="", ylab="")
rect(4.5, 1, 5.5, 3)
rect(c(1, 7.5), c(6, 6), c(2.5, 9), c(8, 8))

}
