% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssHelpers.R
\name{stopAtLine}
\alias{stopAtLine}
\title{trigger a stop, with a message}
\usage{
stopAtLine(msg, line.number)
}
\arguments{
\item{msg}{character, message to display}

\item{line.number}{integer, used as part of the error message}
}
\description{
trigger a stop, with a message
}
\keyword{internal}
