% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{stripchart}
\alias{stripchart}
\title{Draw styled strip chart}
\usage{
stripchart(x, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x}{list of numeric vectors}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of graphics::stripchart}
}
\description{
This is a wrapper for R's stripchart function.
See R's documentation for graphics::stripchart for further details.
}
