% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{svg}
\alias{svg}
\title{Create a styled svg figure}
\usage{
svg(filename, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{filename}{character string with file name}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{character, style class}

\item{...}{Further parameters, see documentation of grDevices::svg}
}
\description{
This is a wrapper for R's svg function.
See R's documentation for grDevices::svg for further details
}
\examples{

# send content of graphics to a pdf file
# to run this, un-comment the pdf() and dev.off() lines
# svg(file="example-file.svg")
barplot(1:5)
# dev.off()

}
