% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{text}
\alias{text}
\title{Add styled text to a plot}
\usage{
text(
  x,
  y = NULL,
  labels = seq_along(x),
  Rcss = "default",
  Rcssclass = NULL,
  ...
)
}
\arguments{
\item{x, y}{coordinates where to write labels}

\item{labels}{characters to print on the plot}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of graphics::text}
}
\description{
This is a wrapper for R's text function.
See R's documentation for graphics::text for further details.
}
\examples{

# add text to an existing plot
plot(c(0, 1), c(0, 1), type="n")
text(0.1, 0.1, "A")
text(c(0.2, 0.7), c(0.8, 0.6), c("B", "C"))

}
