% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{title}
\alias{title}
\title{Add styled annotation to a plot}
\usage{
title(
  main = NULL,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  Rcss = "default",
  Rcssclass = NULL,
  ...
)
}
\arguments{
\item{main}{plot title}

\item{sub}{plot sub title}

\item{xlab, ylab}{labels on axes}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of graphics::title}
}
\description{
This is a wrapper for R's title function.
See R's documentation for graphics::title for further details.
}
\examples{

# add a title
plot(c(0, 1), c(0, 1), type="n", xlab="", ylab="")
title("This is the title")
title(sub="This is a bottom title")

}
