% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{hist}
\alias{hist}
\title{#' Draw a styled histogram}
\usage{
hist(x, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x}{numeric vector}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of graphics::hist}
}
\description{
This is a wrapper for R's hist function.
See R's documentation for graphics::hist for further details.
}
