% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{jpeg}
\alias{jpeg}
\title{Create a styled jpg figure}
\usage{
jpeg(file, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{file}{character string with file name}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of grDevices::jpeg}
}
\description{
This is a wrapper for R's jpeg function.
See R's documentation for grDevices::jpeg for further details
}
\examples{

# send content of graphics to a jpg file
# to run this, un-comment the jpeg() and dev.off() lines
# jpeg(file="example-file.jpg")
barplot(1:5)
# dev.off()

}
