% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{legend}
\alias{legend}
\title{Add a styled legend to aplot}
\usage{
legend(x, y = NULL, legend, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x, y}{position of the legend}

\item{legend}{character vector with labels (text appears in the legend)}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of graphics::legend}
}
\description{
This is a wrapper for R's legend function.
See R's documentation for graphics::legend for further details.
}
