% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{mtext}
\alias{mtext}
\title{Write styled text into a plot margin}
\usage{
mtext(text, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{text}{characters to print on the plot}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of graphics::mtext}
}
\description{
This is a wrapper for R's mtext function.
See R's documentation for graphics::mtext for further details.
}
