% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{par}
\alias{par}
\title{Set styled parameters for base graphics}
\usage{
par(Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of graphics::par}
}
\description{
This is a wrapper for R's par function.
See R's documentation for graphics::par for further details.
}
