% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{points}
\alias{points}
\title{Add styled points to a plot}
\usage{
points(x, y = NULL, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x, y}{coordinates for points on the plot}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{character, style class}

\item{...}{Further parameters, see documentation of graphics::points}
}
\description{
This is a wrapper for R's points function.
See R's documentation for graphics::points for further details.
}
\examples{

# draw a set of points onto an existing plot
plot(c(0, 1), c(0, 1), type="n")
points(runif(10), runif(10))
points(runif(10), runif(10), col="blue", pch=19)

}
