% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{Rcss}
\alias{Rcss}
\title{Create an Rcss style object}
\usage{
Rcss(file = NULL, default = FALSE, overload = FALSE)
}
\arguments{
\item{file}{filename containing Rcss definitions. If set to NULL,
function returns a basic Rcss object. If multiple files, function
reads each one and produces a joint style.}

\item{default}{logical. Should the style object be used as a default
style? See also RcssSetDefaultStyle()}

\item{overload}{logical. Should the standard graphics functions be
overloaded by their Rcss wrappers? See also RcssOverload()}
}
\description{
Creates a style sheet object using definition specified in an
Rcss file. When a file is not specified, creates a base object
object without any styling.
}
\details{
See also related functions RcssSetDefaultStyle() and RcssOverload().
}

