% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rcss.R
\name{Rcss}
\alias{Rcss}
\title{Create an Rcss style object}
\usage{
Rcss(file = NULL, default = FALSE)
}
\arguments{
\item{file}{filename containing Rcss definitions. If set to NULL,
function returns a basic Rcss object.}

\item{default}{(logical) Should the style object be used as a default
style ? See also RcssGetDefaultStyle() and RcssSetDefaultStyle()}
}
\description{
Creates a style sheet object using definition specified in an
Rcss file. When a file is not specified, creates a base object
object without any styling.
}
\details{
See also related functions RcssGetDefaultStyle() and RcssSetDefaultStyle()
}

