% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{Rcss}
\alias{Rcss}
\title{Create an Rcss style object}
\usage{
Rcss(file = NULL)
}
\arguments{
\item{file}{filename containing Rcss definitions. If set to NULL,
function returns a basic Rcss object. If multiple files, function
reads each one and produces a joint style.}
}
\description{
Creates a style sheet object using definition specified in an
Rcss file. When a file is not specified, creates a base object
object without any styling.
}
\details{
See also related functions RcssGetDefaultStyle() and RcssOverload().
}

