% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcssaxis}
\alias{Rcssaxis}
\title{Add a styled axis to a plot}
\usage{
Rcssaxis(side, Rcss = NULL, Rcssclass = NULL, ...)
}
\arguments{
\item{side}{integer specifying what side of the plot to draw the axis.
The codes are 1: bottom, 2: left, 3: top, 4: top.
vertices}

\item{Rcss}{style sheet object}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of axis()}
}
\description{
Rcssaxis is a wrapper for R's axis() function.
See R's documentation for axis() for further details.
}

