% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssGetCompulsoryClass}
\alias{RcssGetCompulsoryClass}
\title{Get current state of compulsory Rcssclass}
\usage{
RcssGetCompulsoryClass(Rcssclass = NULL)
}
\arguments{
\item{Rcssclass}{character vector, set of additional compulsory classes.
When NULL, function returns the current set of compulsory classes
defined in parent environments. When non-NULL, functions returns
the concatentation of the current set and new set.}
}
\description{
Fetches the value of the RcssCompulsoryClass object defined in
parent environments.
}

