% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssGetDefaultStyle}
\alias{RcssGetDefaultStyle}
\title{Get current default Rcssplot style sheet}
\usage{
RcssGetDefaultStyle()
}
\description{
Get a copy of the current default Rcssplot style sheet. See also
related RcssSetDefaultStyle()
}

