% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssGet.R
\name{RcssGetPropertyValue}
\alias{RcssGetPropertyValue}
\title{Extract a value for an Rcss property}
\usage{
RcssGetPropertyValue(Rcss, selector, property, Rcssclass = NULL)
}
\arguments{
\item{Rcss}{style sheet object}

\item{selector}{name of selector of interest (e.g. "plot", "axis",
"text", etc.)}

\item{property}{name of property of interest (e.g. "col", "pch", etc.)}

\item{Rcssclass}{subclass of style sheet}
}
\description{
Extract a value for a property from an Rcss style sheet object. Returns
a list with two items. "Defined" is a boolean that indicates the
property is defined in the style sheet. "Value" gives the actual value
of the property.
}
\details{
Equivalent to RcssProperty; use RcssProperty instead.
}
