% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssLexer.R
\name{RcssLexChars}
\alias{RcssLexChars}
\title{The root of the actual lexer.
Technically, this keeps track of the lexed tokens.
Dispatches to other tokens to parse the next tokens.}
\usage{
RcssLexChars(cc, pos = 1)
}
\arguments{
\item{cc}{a vector of characters}

\item{position}{current position in the character vector}

\item{terminals}{vector with characters for terminal characters}
}
\value{
data frame with pairs (token, tokentype)
}
\description{
The root of the actual lexer.
Technically, this keeps track of the lexed tokens.
Dispatches to other tokens to parse the next tokens.
}
