% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssLexer.R
\name{RcssLexer}
\alias{RcssLexer}
\title{start the Lexer}
\usage{
RcssLexer(f)
}
\arguments{
\item{f}{character, input file, or vector of input files}
}
\value{
data frame with tokens
}
\description{
start the Lexer
}
