% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssOverload}
\alias{RcssOverload}
\title{Overloads base graphics functions by their Rcssplot wrappers}
\usage{
RcssOverload()
}
\description{
Rcssplot graphics functions have 'Rcss' prefixes,
e.g Rcsstext(). This function can be invoked to overload 
base-graphics functions by their Rcss wrappers. i.e. After executing
this function, you can execute e.g. text() and
automatically use the Rcss capabilities.
}
\details{
Warning: this function creates masking objects in your current
environment for many base-graphics functions. See documentation
for details.
}

