% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssLexer.R
\name{RcssParseComment}
\alias{RcssParseComment}
\title{parse a comment}
\usage{
RcssParseComment(chardata, pos)
}
\arguments{
\item{chardata}{data frame with characters and line numbers}

\item{pos}{current position (expects '/' followed by '*')}
}
\value{
position of first non-comment
}
\description{
parse a comment
}
\keyword{internal}
