% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssParser.R
\name{RcssParseDeclaration}
\alias{RcssParseDeclaration}
\title{Parse one declaration (expects an IDENT)}
\usage{
RcssParseDeclaration(lextab, n)
}
\arguments{
\item{lexta}{data frame}
}
\description{
Parse one declaration (expects an IDENT)
}
