% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssLexer.R
\name{RcssParseGeneric}
\alias{RcssParseGeneric}
\title{parser for a generic token}
\usage{
RcssParseGeneric(chardata, pos, delimiters)
}
\arguments{
\item{chardata}{data frame with characters and line numbers}

\item{pos}{integer current position}

\item{delimiters}{vector or delimiters that demarcate end of a token
(e.g. terminals not allowed in a variable name)}
}
\value{
position of first character beyond the current token
}
\description{
parser for a generic token
}
\keyword{internal}
