% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssLexer.R
\name{RcssParseHexToken}
\alias{RcssParseHexToken}
\title{parse a hex color}
\usage{
RcssParseHexToken(cc, pos)
}
\arguments{
\item{cc}{vector of characters}

\item{pos}{integer, current position (this function expects a hash sign)}
}
\value{
position of the next non-hex character
}
\description{
parse a hex color
}
