% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssLexer.R
\name{RcssParseString}
\alias{RcssParseString}
\title{parse a string}
\usage{
RcssParseString(chardata, pos, delimiter = "\\"")
}
\arguments{
\item{chardata}{data frame with characters and line numbers}

\item{pos}{integer, current position (expects '/' followed by '*')}

\item{delimiter}{character, either " or ' (used to catch nested strings)}
}
\value{
- position of first character outside the string
}
\description{
parse a string
}
\keyword{internal}
