% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcssWrappers.R
\name{Rcsspolygon}
\alias{Rcsspolygon}
\title{Draw a styled polygon on a plot}
\usage{
Rcsspolygon(x, y = NULL, Rcss = "default", Rcssclass = NULL, ...)
}
\arguments{
\item{x,y}{coordinates for polygon vertices}

\item{Rcss}{style sheet object. Leave "default" to use a style
defined via RcssSetDefaultStyle()}

\item{Rcssclass}{sub class of style sheet}

\item{...}{Further parameters, see documentation of polygon()}
}
\description{
Rcsspolygon is a wrapper for R's polygon() function.
See R's documentation for polygon() for further details.
}

