% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssGet.R
\name{RcssProperty}
\alias{RcssProperty}
\title{Extract information about property and its value}
\usage{
RcssProperty(selector, property, Rcssclass = NULL, Rcss = "default")
}
\arguments{
\item{selector}{character, name of selector, e.g. 'points'}

\item{property}{character, name of property, e.g. 'col'}

\item{Rcssclass}{character or vector, subclass in Rcss}

\item{Rcss}{Rcss object}
}
\description{
See also RcssGetPropertyValue.
}
