% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssSetDefaultStyle}
\alias{RcssSetDefaultStyle}
\title{Set default Rcssplot style sheet}
\usage{
RcssSetDefaultStyle(Rcss)
}
\arguments{
\item{Rcss}{style sheet object (set NULL to remove default style)}
}
\description{
Set a default Rcssplot style sheet. This style sheet will be
applied in all functions of the Rcss family. See also related
function RcssGetDefaultStyle()
}

