% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcss.R
\name{RcssSetDefaultStyle}
\alias{RcssSetDefaultStyle}
\title{Set default Rcssplot style sheet}
\usage{
RcssSetDefaultStyle(Rcss, overload = FALSE)
}
\arguments{
\item{Rcss}{style sheet object (set NULL to remove default style)}

\item{overload}{logical. Should base graphics function be overloaded
by the Rcssplot wrappers?}
}
\description{
Set a default Rcssplot style sheet. This style sheet will be
applied in all functions of the Rcss family. See also related
function RcssGetDefaultStyle()
}

