% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcssWatch.R
\name{RcssWatch}
\alias{RcssWatch}
\title{development tool for adjusting Rcss and R graphics code}
\usage{
RcssWatch(f, Rcss.files = NULL, R.files = NULL, ...)
}
\arguments{
\item{f}{function or character of function name, executed at each iteration}

\item{Rcss.files}{character, paths to Rcss files}

\item{R.files}{character, paths to R files}

\item{...}{other arguments, passed to function f}
}
\description{
This is a macro script that loads R code and a default Rcss style,
and then executes a function. This process is repeated indefinitely.
}
\examples{

# this repeatedly draws and redraws a bar plot
RcssWatch(plot, x=1:4, y=1:4)
# for more interesting behavior, specify a files with styles and R source

}
